requirements.txt
numpy>=1.18.1
scipy>=1.4.1
#tensorflow==2.1.0
#astropy>=4.0
#matplolib>=3.1.2
celluloid
#tensorflow-addons~=0.7
tensorflow-gpu
astropy
#poppy==0.9.0
matplotlib
Pillow>=7.0.0
scikit-learn==0.23.1
